create-obs-discharge-csv/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
attrs=21.2.0=pyhd8ed1ab_0
boost-cpp=1.74.0=h5b4e17d_4
bzip2=1.0.8=h8ffe710_4
ca-certificates=2021.5.30=h5b45459_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cairo=1.16.0=hb19e0ff_1008
certifi=2021.5.30=py39hcbf5309_0
cfitsio=3.470=h0af3d06_7
cftime=1.5.0=py39h5d4886f_0
click=7.1.2=pyh9f0ad1d_0
click-plugins=1.1.1=py_0
cligj=0.7.2=pyhd8ed1ab_0
curl=7.78.0=h789b8ee_0
cycler=0.10.0=py_2
decorator=4.4.2=py_0
expat=2.4.1=h39d44d4_0
fiona=1.8.20=py39hea8b339_1
fontconfig=2.13.1=h1989441_1005
freetype=2.10.4=h546665d_1
freexl=1.0.6=ha8e266a_0
gdal=3.3.1=py39h6795fcd_1
geopandas=0.9.0=pyhd8ed1ab_1
geopandas-base=0.9.0=pyhd8ed1ab_1
geos=3.9.1=h39d44d4_2
geotiff=1.6.0=ha8a8a2d_6
gettext=0.19.8.1=h1a89ca6_1005
h5py=3.3.0=nompi_py39hf27771d_100
hdf4=4.2.15=h0e5069d_3
hdf5=1.10.6=nompi_h5268f04_1114
icu=68.1=h0e60522_0
intel-openmp=2021.3.0=h57928b3_3372
jbig=2.1=h8d14728_2003
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h8ffe710_0
kealib=1.4.14=h96bfa42_2
kiwisolver=1.3.1=py39h2e07f2f_1
krb5=1.19.2=hbae68bd_0
lcms2=2.12=h2a16943_0
lerc=2.2.1=h0e60522_0
libblas=3.9.0=10_mkl
libcblas=3.9.0=10_mkl
libclang=11.1.0=default_h5c34c98_1
libcurl=7.78.0=h789b8ee_0
libdeflate=1.7=h8ffe710_5
libffi=3.3=h0e60522_2
libgdal=3.3.1=ha6cada3_1
libglib=2.68.3=h1e62bf3_0
libiconv=1.16=he774522_0
libkml=1.3.0=h9859afa_1014
liblapack=3.9.0=10_mkl
libnetcdf=4.8.0=nompi_hf689e7d_103
libpng=1.6.37=h1d00b33_2
libpq=13.3=hfcc5ef8_0
librttopo=1.1.0=hb340de5_6
libspatialindex=1.9.3=h39d44d4_4
libspatialite=5.0.1=h762a7f4_5
libssh2=1.9.0=h680486a_6
libtiff=4.3.0=h0c97f57_1
libwebp-base=1.2.0=h8ffe710_2
libxml2=2.9.12=hf5bbc77_0
libzip=1.8.0=hfed4ece_0
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
mapclassify=2.4.3=pyhd8ed1ab_0
matplotlib=3.4.2=py39hcbf5309_0
matplotlib-base=3.4.2=py39h581301d_0
mkl=2021.3.0=hb70f87d_564
msys2-conda-epoch=20160418=1
munch=2.5.0=py_0
netcdf4=1.5.7=nompi_py39hc20e565_100
networkx=2.6.2=pyhd8ed1ab_0
numpy=1.21.1=py39h6635163_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=hb211442_1
openssl=1.1.1k=h8ffe710_0
pandas=1.3.1=py39h2e25243_0
pcraster=4.3.1=py39hd42c513_1
pcre=8.45=h0e60522_0
pillow=8.3.1=py39h916092e_0
pip=21.2.2=pyhd8ed1ab_0
pixman=0.40.0=h8ffe710_0
poppler=21.03.0=h9ff6ed8_0
poppler-data=0.4.10=0
postgresql=13.3=h1c22c4f_0
proj=8.0.1=h1cfcee9_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyproj=3.1.0=py39ha996c60_3
pyqt=5.12.3=py39hcbf5309_7
pyqt-impl=5.12.3=py39h415ef7b_7
pyqt5-sip=4.19.18=py39h415ef7b_7
pyqtchart=5.12=py39h415ef7b_7
pyqtwebengine=5.12.1=py39h415ef7b_7
pyreadline=2.1=py39hcbf5309_1003
python=3.9.6=h7840368_1_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2021.1=pyhd8ed1ab_0
qt=5.12.9=h5909a2a_4
rtree=0.9.7=py39h09fdee3_2
scikit-learn=0.24.2=py39he931e04_0
scipy=1.7.1=py39hc0c34ad_0
setuptools=49.6.0=py39hcbf5309_3
shapely=1.7.1=py39haadaec5_5
six=1.16.0=pyh6c4a22f_0
sqlite=3.36.0=h8ffe710_0
tbb=2021.3.0=h2d74725_0
threadpoolctl=2.2.0=pyh8a188c0_0
tiledb=2.3.2=h78dabda_0
tk=8.6.10=h8ffe710_1
tornado=6.1=py39hb82d6ee_1
tzdata=2021a=he74cb21_1
ucrt=10.0.20348.0=h57928b3_0
vc=14.2=hb210afc_5
vs2015_runtime=14.29.30037=h902a5da_5
wheel=0.36.2=pyhd3deb0d_0
wincertstore=0.2=py39hcbf5309_1006
xerces-c=3.2.3=h0e60522_2
xz=5.2.5=h62dcd97_1
zlib=1.2.11=h62dcd97_1010
zstd=1.5.0=h6255e5f_0
